modified:   tt2/Flow.p
We have signs of a chainsetups. Initializations are
screwed up. The sequence was always fragile.

The only way to sequence them is as objects.
The "program" doesn't initiaziae.

Of course you have this already in yourprocedural code.

so find which places have gui which subroutines

go in a new branch called v_obj

this will serve as your test bed