modified:   tt2/Flow.p
YAML header not found. Today I am going
to write tests as I bring up what I need.

for example strip_comments.

It is much easier at this point to test this
way, than to rely on debugging statements.

So, another laborsaving innovation.

I'll put this into the chain, so that
it happens at make test.