On branch refs/heads/v_91
	new file:   UI/pre/UI.p.test

This is the test version.

	modified:   UI/t/1.t

This is the latest test, with the right
libraries added (and a bunch more).

My next project is that there are
subroutine calls in UI::Graphical
that mean to get at subs in UI.

There is no inheritance without the
object mechanism, so I'll put in lines
like this:

package UI::Text;
sub load_session( UI::load_session() );
This is cheap and will fix that problem.