git_snapshot() incorporates save_state()
I'm considering if a special attribute in commands.yml is needed for
commands that reset the buffer used for commit messages.

It is mistakenly called undo_buffer, we will rename it
command_context_buffer, since it holds the commands
and context for a git commit.

A buffer may not be necessary in the autosave mode
but we already have one and its cool.

I note that that context is not complete, as we have some
new pronouns, notable $this_sequence, which is like
$this_bus, except it only ever contains sequences or undef.