add Effect attributes: nick, surname nick is a name attribute of an effect. It is not unique. Several tracks may have a "reverb" effect.
when the user defines a nick, she also
defines an effect chain of the same
name, which becomes a preset.

If the user wishes to always apply
the default settings, she can "nick"
the effect before modifying any parameters.

"nick overwrite" could be syntax for
replacing a nick.

"surname" is intended to be set by
multieffect effect chain, when it would
be confusing to have mulitiple tracks
by the same name, it makes sense that they
could all share a surname.

This will allow us to support rfx reverb_chain.